# sql: use new batch limit to limit all scans