We change the kvFetcher code to use the new per-batch limit instead of limiting
only single-span scans.

The scan_test code is modified to work with multiple spans (previous code
assumed OR branches would result in spans but that's not the case)

The benchmarks use a table with 10 * 1000 rows. The query involves 4 spans of
1000 rows each, with the filter passing on only the first 10,30,50,70 rows
respectively.

```
name                                old time/op  new time/op  delta
Scan10000FilterLimit1_Cockroach-4   10.1ms ±51%   0.3ms ±23%  -97.43%  (p=0.000 n=10+9)
Scan10000FilterLimit10_Cockroach-4  9.07ms ±17%  0.27ms ± 2%  -97.05%   (p=0.000 n=9+9)
Scan10000FilterLimit50_Cockroach-4  10.5ms ± 8%  10.8ms ± 9%     ~     (p=0.211 n=10+9)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="35" align="absmiddle" alt="Review on Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/5214)
<!-- Reviewable:end -->