@petermattis 2016-03-15
Why does the `Scan10000FilterLimit50_Cockroach` benchmark not show improvement from this change?

<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/> 

---

Reviewed 1 of 1 files at r1.
Review status: 1 of 3 files reviewed at latest revision, all discussions resolved.

---



*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/5214#-:-KCusnPShxnlPEtPnzHz:1576193991)*
<!-- Sent from Reviewable.io -->