@RaduBerinde 2016-03-15
We will retrieve 50 rows but that will not be enough to return 50 results because of the `b < 10*a` filter. We then continue fetching with the regular batch size. There are various things we can improve here (multiply the initial batch size by a factor depending on filter, or grow the batch size more slowly, e.g. double it each time) , but this is expected for now.

---

Review status: 1 of 3 files reviewed at latest revision, all discussions resolved, all commit checks successful.

---



*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/5214#-:-KCuu0irD6F9WQIZynKf:-1923038733)*
<!-- Sent from Reviewable.io -->