@petermattis 2016-03-15
Ah, that makes sense.

---

Review status: 1 of 3 files reviewed at latest revision, all discussions resolved, all commit checks successful.

---



*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/5214#-:-KCv-0uAQ65sqIY2Zipn:-1842318892)*
<!-- Sent from Reviewable.io -->